move git files to git package
