+++ git/repo_parser_test.go
@@ -1,4 +1,4 @@
-package utils
+package git
 
 import (
 	"errors"
